+++ 01_ecli/src/main.rs
@@ -1,30 +1,37 @@
-use std::{fs, path::Path};
-
 use anyhow::Result;
+use serde::Deserialize;
+use std::fs::File;
 
-// 当不需要传递错误信息时，可以使用 `Result<(), Error>` 作为返回值
-fn main() -> Result<()> {
-    let content = read_file("assets/juventus.csv")?;
-
-    println!("{}", content);
-    Ok(())
+#[allow(dead_code)]
+#[derive(Debug, Deserialize)]
+struct Player {
+    // #[serde(rename = "Name")] 属性用于将 CSV 列名映射为结构字段名。
+    #[serde(rename = "Name")]
+    name: String,
+    #[serde(rename = "Position")]
+    position: String,
+    #[serde(rename = "DOB")]
+    dob: String,
+    #[serde(rename = "Nationality")]
+    nationality: String,
+    #[serde(rename = "Kit Number")]
+    number: u8,
 }
 
-// 适合在需要传递错误信息时使用 `Result<String, Error>` 作为返回值
-// fn main() {
-//     let content = read_file("assets/juventus.csv");
-//
-//     match content {
-//         Ok(c) => println!("Content: {}", c),
-//         Err(e) => println!("Error: {}", e),
-//     }
-// }
+fn main() -> Result<()> {
+    let file = File::open("assets/juventus.csv")?;
+    let mut reader = csv::ReaderBuilder::new()
+        .has_headers(true)
+        .from_reader(file);
 
-fn read_file<P>(file_path: P) -> Result<String>
-where
-    P: AsRef<Path>,
-{
-    let content = fs::read_to_string(file_path)?; // err convert to Result
+    for result in reader.deserialize() {
+        let player: Player = result?;
 
-    Ok(content)
+        println!(
+            "Name: {}, Position: {}, DOB: {}, Nationality: {}, Number: {}",
+            player.name, player.position, player.dob, player.nationality, player.number,
+        );
+    }
+
+    Ok(())
 }